+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # stripe-angular - Change Log
 All notable changes to this project will be documented here.
 
+## [0.4.1] - (2019-03-25)
+- Made stripe-source a base component of stripe-bank and stripe-card
+
 ## [0.4.0] - (2019-03-25)
-- Added support for creating sources instead of only tokens
+- BETA: Added support for creating sources as stripe-source component
 
 ## [0.3.2] - (2018-03-25)
 - Fix loading Stripe when its already been loaded before